:hammer: Merge TileManager and TileManagerBase
